+++ classroom/dataset/__init__.py
@@ -1,10 +1,9 @@
-from .bytesdataset import BytesDataset
 from .utf8 import utf8encode
 from .utf8 import utf8decode
 from .utf8 import utf8bitsencode
 from .utf8 import utf8bitsdecode
-from .randomtokensdataset import RandomTokensDataset
-from .gutenbergsnippetsdataset import GutenbergSnippetsDataset
+from .gpt2 import gpt2encode
+from .gpt2 import gpt2decode
 from .gutenbergbitsdataset import GutenbergBitsDataset
 from .gutenbergbytesdataset import GutenbergBytesDataset
-from .gutenberggpt2dataset import GutenbergGPT2Dataset, gpt2decode, gpt2encode
+from .gutenberggpt2dataset import GutenbergGPT2Dataset
