+++ classroom/gui/__init__.py
@@ -1,3 +1,11 @@
 from bokeh.io import output_notebook
 output_notebook()
 from .plot import Plot
+from .fun import Fun
+from .filters import Count
+from .filters import Diff
+from .filters import Sum
+from .filters import Log2Sum
+from .filters import KalmanFilter1D
+from .filters import MedianFilter
+from .filters import TwoWindowFilter
