+++ classroom/trainer/__init__.py
@@ -81,49 +81,3 @@ class Trainer:
         loss = self.optimizer.step(closure)
         self.n += 1
         return loss, baseline_loss
-
-    @torch.no_grad()
-    def autocomplete(self, prompt=None, n_generate=128, n_ctx=None, temp=1.0, encode=None, decode=None, output=None):
-        """
-        Autocomplete using the model
-
-        ## Args
-        * `prompt: str` an optional prompt to begin with
-        * `n_generate: int` the number of bytes/tokens to generate
-        * `n_ctx: int` the number of bytes/tokens in the context window
-        * `encode` the function that can turn an str into a sequence of bytes/tokens suitable for the model.
-        defaults to utf8encode
-        * `decode` the function that can turn the sequences of bytes/tokens used by the model to a str
-        defaults to utf8decode
-        * `output: Optional[List[int]]` a list to stream the output bytes/tokens to (as `int`s; they will not be decoded to `str`).
-
-        ## TODO
-        * make streaming autocomplete with streamed characters (i.e. length 1 strings) using asyncio
-        """
-        Categorical = torch.distributions.Categorical
-        if n_ctx is None:
-            n_ctx = self.model.n_ctx
-        if encode is None:
-            encode = self.dataset.encode # utf8encode, gpt2encode
-        if decode is None:
-            decode = self.dataset.decode # utf8decode, gpt2decode
-        if prompt is None:
-            prompt = decode(self.dataset.batch(1, 2*n_ctx, offset=None).tolist()[0])  # kludge
-        x = encode(prompt)
-        x = x[-n_ctx:]
-        prompt = decode(x)
-        print(f"=== Prompt ===\n{prompt}\n=== Autocompletion ===\n")
-
-        def sampler(x):
-            x = list(x)
-            for _ in range(n_generate):
-                probs = self.model.inference(torch.tensor(x, dtype=torch.long, device=self.device).unsqueeze(0)).view(-1)[-self.model.n_vocab_out:]
-                if temp > 0:
-                    y = Categorical(probs=probs**(1.0/temp)).sample().item()
-                else:
-                    y = torch.argmax(probs).item()
-                x = (x + [y])[-n_ctx:]
-                if output is not None:
-                    output.append(y)
-                yield y
-        return decode(list(sampler(x)))
