--- classroom/util/__init__.py
@@ -1,8 +1,0 @@
-from .fun import Fun
-from .filters import Count
-from .filters import Diff
-from .filters import Sum
-from .filters import Log2Sum
-from .filters import KalmanFilter1D
-from .filters import MedianFilter
-from .filters import TwoWindowFilter
